requirements.txt
torch>=2.0
torchvision>=0.15
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
pillow>=9.5
numpy>=1.24,<2.0
jupyterlab>=4.0
opencv-python>=4.0.0
torchinfo>=1.7